backend/requirements.txt
alembic==1.17.2
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
greenlet==3.3.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
gunicorn==23.0.0
psycopg2-binary==2.9.9
